+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/CheckMojo.java
@@ -1,6 +1,7 @@
 package nbbrd.heylogs.maven.plugin;
 
-import internal.heylogs.StylishFormat;
+import nbbrd.console.picocli.MultiFileInputOptions;
+import nbbrd.console.picocli.text.TextOutputSupport;
 import nbbrd.heylogs.Check;
 import nbbrd.heylogs.Heylogs;
 import nbbrd.heylogs.ext.semver.SemVer;
@@ -13,46 +14,79 @@ import org.apache.maven.plugins.annotations.Parameter;
 import java.io.File;
 import java.io.IOException;
 import java.io.Writer;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
 
-import static internal.heylogs.maven.plugin.HeylogsParameters.*;
-import static java.util.Collections.singletonList;
+import static internal.heylogs.FormatSupport.resolveFormatId;
+import static internal.heylogs.HeylogsParameters.*;
 import static java.util.Locale.ROOT;
+import static java.util.stream.Collectors.toList;
+import static nbbrd.console.picocli.ByteOutputSupport.DEFAULT_STDOUT_FILE;
+import static nbbrd.console.picocli.text.TextOutputSupport.newTextOutputSupport;
 
+@lombok.Getter
+@lombok.Setter
 @Mojo(name = "check", defaultPhase = LifecyclePhase.VALIDATE, threadSafe = true, requiresProject = false)
 public final class CheckMojo extends HeylogsMojo {
 
-    @Parameter(defaultValue = WORKING_DIR_CHANGELOG, property = INPUT_FILE_PROPERTY)
-    private File inputFile;
+    @Parameter(property = "heylogs.inputFiles", defaultValue = DEFAULT_CHANGELOG_FILE)
+    private List<File> inputFiles;
 
-    @Parameter(defaultValue = MOJO_LOG_FILE, property = OUTPUT_FILE_PROPERTY)
+    @Parameter(property = "heylogs.recursive", defaultValue = DEFAULT_RECURSIVE)
+    private boolean recursive;
+
+    @Parameter(property = "heylogs.outputFile", defaultValue = DEFAULT_STDOUT_FILE)
     private File outputFile;
 
-    @Parameter(defaultValue = "false", property = "heylogs.semver")
+    @Parameter(property = "heylogs.semver", defaultValue = DEFAULT_SEMVER)
     private boolean semver;
 
-    @Parameter(defaultValue = StylishFormat.ID, property = FORMAT_ID_PROPERTY)
+    @Parameter(property = "heylogs.formatId")
     private String formatId;
 
     @Override
     public void execute() throws MojoExecutionException {
-        if (skip) {
+        if (isSkip()) {
             getLog().info("Checking has been skipped.");
             return;
         }
 
         if (semver) {
-            checkVersioning(new SemVer(), getProjectVersionOrNull());
+            checkVersioning(new SemVer(), getProjectVersion());
         }
 
-        if (inputFile.exists()) {
-            check();
-        } else {
-            if (isRootProject()) {
-                raiseErrorMissingChangelog();
-            } else {
-                notifyMissingChangelog();
+        Heylogs heylogs = initHeylogs(semver);
+
+        List<Check> list = new ArrayList<>();
+        try {
+            MultiFileInputOptions input = new MultiFileInputOptions();
+            input.setFiles(inputFiles.stream().map(File::toPath).collect(toList()));
+            input.setRecursive(recursive);
+            input.setSkipErrors(false);
+            for (Path file : input.getAllFiles(HeylogsMojo::accept)) {
+                list.add(Check
+                        .builder()
+                        .source(file.toString())
+                        .problems(heylogs.checkFormat(readChangelog(file.toFile())))
+                        .build());
             }
+        } catch (IOException ex) {
+            throw new MojoExecutionException("Error while listing files", ex);
         }
+        boolean hasErrors = list.stream().anyMatch(Check::hasErrors);
+
+        TextOutputSupport outputSupport = newTextOutputSupport();
+        String formatId = resolveFormatId(getFormatId(), heylogs, outputSupport::isStdoutFile, outputFile.toPath());
+
+        try (Writer writer = newWriter(outputFile, hasErrors ? getLog()::error : getLog()::info)) {
+            heylogs.formatProblems(formatId, writer, list);
+        } catch (IOException ex) {
+            throw new MojoExecutionException("Error while writing", ex);
+        }
+
+        if (hasErrors)
+            throw new MojoExecutionException("Invalid changelog");
     }
 
     private void checkVersioning(Versioning versioning, String projectVersion) throws MojoExecutionException {
@@ -64,26 +98,6 @@ public final class CheckMojo extends HeylogsMojo {
         } else {
             getLog().error(String.format(ROOT, "Invalid project version: '%s' must follow %s", projectVersion, versioning.getVersioningName()));
             throw new MojoExecutionException("Invalid project version. See above for details.");
-        }
-    }
-
-    private void check() throws MojoExecutionException {
-        Heylogs heylogs = initHeylogs(semver);
-
-        Check check = Check
-                .builder()
-                .source(inputFile.toString())
-                .problems(heylogs.checkFormat(readChangelog(inputFile)))
-                .build();
-
-        try (Writer writer = newWriter(outputFile, check.hasErrors() ? getLog()::error : getLog()::info)) {
-            heylogs.formatProblems(formatId, writer, singletonList(check));
-        } catch (IOException ex) {
-            throw new MojoExecutionException("Error while writing", ex);
-        }
-
-        if (check.hasErrors()) {
-            throw new MojoExecutionException("Invalid changelog");
         }
     }
 }