+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/ExtractMojo.java
@@ -1,53 +1,58 @@
 package nbbrd.heylogs.maven.plugin;
 
 import com.vladsch.flexmark.util.ast.Document;
-import internal.heylogs.maven.plugin.MojoFunction;
+import internal.heylogs.maven.plugin.MojoParameterParsing;
+import lombok.NonNull;
 import nbbrd.heylogs.Filter;
-import nbbrd.heylogs.Heylogs;
 import nbbrd.heylogs.TimeRange;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.File;
 import java.time.LocalDate;
+import java.time.format.DateTimeParseException;
 import java.util.Objects;
 import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
 
-import static internal.heylogs.maven.plugin.HeylogsParameters.*;
-import static internal.heylogs.maven.plugin.MojoFunction.of;
+import static internal.heylogs.HeylogsParameters.DEFAULT_CHANGELOG_FILE;
+import static nbbrd.console.picocli.ByteOutputSupport.DEFAULT_STDOUT_FILE;
 
+@lombok.Getter
+@lombok.Setter
 @Mojo(name = "extract", defaultPhase = LifecyclePhase.GENERATE_RESOURCES, threadSafe = true, requiresProject = false)
 public final class ExtractMojo extends HeylogsMojo {
 
-    @Parameter(defaultValue = WORKING_DIR_CHANGELOG, property = INPUT_FILE_PROPERTY)
+    @Parameter(property = "heylogs.inputFile", defaultValue = DEFAULT_CHANGELOG_FILE)
     private File inputFile;
 
-    @Parameter(defaultValue = "${project.build.directory}/CHANGELOG.md", property = OUTPUT_FILE_PROPERTY)
+    @Parameter(property = "heylogs.outputFile", defaultValue = DEFAULT_STDOUT_FILE)
     private File outputFile;
 
-    @Parameter(defaultValue = "${project.version}", property = "heylogs.ref")
+    @Parameter(property = "heylogs.ref", defaultValue = "${project.version}")
     private String ref;
 
-    @Parameter(defaultValue = "-999999999-01-01", property = "heylogs.from")
+    @Parameter(property = "heylogs.from", defaultValue = "-999999999-01-01")
     private String from;
 
-    @Parameter(defaultValue = "+999999999-12-31", property = "heylogs.to")
+    @Parameter(property = "heylogs.to", defaultValue = "+999999999-12-31")
     private String to;
 
-    @Parameter(defaultValue = "0x7fffffff", property = "heylogs.limit")
+    @Parameter(property = "heylogs.limit", defaultValue = "0x7fffffff")
     private int limit;
 
-    @Parameter(defaultValue = "^.*-SNAPSHOT$", property = "heylogs.unreleased.pattern")
+    @Parameter(property = "heylogs.unreleasedPattern", defaultValue = "^.*-SNAPSHOT$")
     private String unreleasedPattern;
 
-    @Parameter(defaultValue = "false", property = "heylogs.ignore.content")
+    @Parameter(property = "heylogs.ignoreContent", defaultValue = "false")
     private boolean ignoreContent;
 
     @Override
     public void execute() throws MojoExecutionException {
-        if (skip) {
+        if (isSkip()) {
             getLog().info("Extracting has been skipped.");
             return;
         }
@@ -57,32 +62,38 @@ public final class ExtractMojo extends HeylogsMojo {
             throw new MojoExecutionException("Changelog not found");
         }
 
-        extract(loadFilter());
-    }
-
-    private Filter loadFilter() throws MojoExecutionException {
-        return Filter
-                .builder()
-                .ref(Objects.toString(ref, ""))
-                .unreleasedPattern(UNRELEASED_PATTERN_PARSER.applyWithMojo(unreleasedPattern))
-                .timeRange(TimeRange.of(FROM_PARSER.applyWithMojo(from), TO_PARSER.applyWithMojo(to)))
-                .limit(limit)
-                .ignoreContent(ignoreContent)
-                .build();
-    }
-
-    private void extract(Filter filter) throws MojoExecutionException {
-        Heylogs heylogs = initHeylogs(false);
-
         Document changelog = readChangelog(inputFile);
 
+        Filter filter = toFilter();
+
         getLog().info("Extracting with " + filter);
-        heylogs.extractVersions(changelog, filter);
+        initHeylogs(false).extractVersions(changelog, filter);
 
         writeChangelog(changelog, outputFile);
     }
 
-    private static final MojoFunction<String, Pattern> UNRELEASED_PATTERN_PARSER = of(Pattern::compile, "Invalid unreleased pattern");
-    private static final MojoFunction<String, LocalDate> FROM_PARSER = of(Filter::parseLocalDate, "Invalid format for 'from' parameter");
-    private static final MojoFunction<String, LocalDate> TO_PARSER = of(Filter::parseLocalDate, "Invalid format for 'to' parameter");
+    @MojoParameterParsing
+    private @NonNull Filter toFilter() throws MojoExecutionException {
+        try {
+            return Filter
+                    .builder()
+                    .ref(Objects.toString(ref, ""))
+                    .unreleasedPattern(Pattern.compile(unreleasedPattern))
+                    .timeRange(TimeRange.of(parseLocalDate(from), parseLocalDate(to)))
+                    .limit(limit)
+                    .ignoreContent(ignoreContent)
+                    .build();
+        } catch (PatternSyntaxException ex) {
+            throw new MojoExecutionException("Invalid unreleased pattern", ex);
+        }
+    }
+
+    private static @NonNull LocalDate parseLocalDate(@Nullable String text) throws MojoExecutionException {
+        if (text == null) throw new MojoExecutionException("Missing date");
+        try {
+            return Filter.parseLocalDate(text);
+        } catch (DateTimeParseException ex) {
+            throw new MojoExecutionException("Invalid date format", ex);
+        }
+    }
 }