+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/ListMojo.java
@@ -1,6 +1,6 @@
 package nbbrd.heylogs.maven.plugin;
 
-import internal.heylogs.StylishFormat;
+import nbbrd.console.picocli.text.TextOutputSupport;
 import nbbrd.heylogs.Heylogs;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
@@ -11,32 +11,36 @@ import java.io.File;
 import java.io.IOException;
 import java.io.Writer;
 
-import static internal.heylogs.maven.plugin.HeylogsParameters.*;
+import static internal.heylogs.FormatSupport.resolveFormatId;
+import static internal.heylogs.HeylogsParameters.DEFAULT_SEMVER;
+import static nbbrd.console.picocli.ByteOutputSupport.DEFAULT_STDOUT_FILE;
+import static nbbrd.console.picocli.text.TextOutputSupport.newTextOutputSupport;
 
+@lombok.Getter
+@lombok.Setter
 @Mojo(name = "list", defaultPhase = LifecyclePhase.VALIDATE, threadSafe = true, requiresProject = false)
 public final class ListMojo extends HeylogsMojo {
 
-    @Parameter(defaultValue = MOJO_LOG_FILE, property = OUTPUT_FILE_PROPERTY)
+    @Parameter(property = "heylogs.outputFile", defaultValue = DEFAULT_STDOUT_FILE)
     private File outputFile;
 
-    @Parameter(defaultValue = "false", property = "heylogs.semver")
+    @Parameter(property = "heylogs.semver", defaultValue = DEFAULT_SEMVER)
     private boolean semver;
 
-    @Parameter(defaultValue = StylishFormat.ID, property = FORMAT_ID_PROPERTY)
+    @Parameter(property = "heylogs.formatId")
     private String formatId;
 
     @Override
     public void execute() throws MojoExecutionException {
-        if (skip) {
+        if (isSkip()) {
             getLog().info("Listing has been skipped.");
             return;
         }
 
-        list();
-    }
-
-    private void list() throws MojoExecutionException {
         Heylogs heylogs = initHeylogs(semver);
+
+        TextOutputSupport outputSupport = newTextOutputSupport();
+        String formatId = resolveFormatId(getFormatId(), heylogs, outputSupport::isStdoutFile, outputFile.toPath());
 
         try (Writer writer = newWriter(outputFile, getLog()::info)) {
             heylogs.formatResources(formatId, writer, heylogs.listResources());