+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/ReleaseMojo.java
@@ -1,44 +1,43 @@
 package nbbrd.heylogs.maven.plugin;
 
 import com.vladsch.flexmark.util.ast.Document;
-import internal.heylogs.maven.plugin.MojoFunction;
-import nbbrd.design.MightBePromoted;
-import nbbrd.heylogs.Heylogs;
+import internal.heylogs.maven.plugin.MojoParameterParsing;
+import lombok.NonNull;
 import nbbrd.heylogs.Version;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.File;
 import java.time.LocalDate;
-import java.time.ZoneId;
+import java.time.format.DateTimeParseException;
 import java.util.Objects;
 
-import static internal.heylogs.maven.plugin.HeylogsParameters.*;
-import static internal.heylogs.maven.plugin.MojoFunction.of;
+import static internal.heylogs.HeylogsParameters.DEFAULT_CHANGELOG_FILE;
+import static nbbrd.console.picocli.ByteOutputSupport.DEFAULT_STDOUT_FILE;
 
+@lombok.Getter
+@lombok.Setter
 @Mojo(name = "release", defaultPhase = LifecyclePhase.GENERATE_RESOURCES, threadSafe = true, requiresProject = false)
 public final class ReleaseMojo extends HeylogsMojo {
 
-    @Parameter(defaultValue = WORKING_DIR_CHANGELOG, property = INPUT_FILE_PROPERTY)
+    @Parameter(property = "heylogs.inputFile", defaultValue = DEFAULT_CHANGELOG_FILE)
     private File inputFile;
 
-    @Parameter(defaultValue = "${project.build.directory}/CHANGELOG.md", property = OUTPUT_FILE_PROPERTY)
-    private File outputFile;
-
-    @Parameter(defaultValue = "${project.version}", property = "heylogs.ref")
+    @Parameter(property = "heylogs.ref", defaultValue = "${project.version}")
     private String ref;
 
-    @Parameter(defaultValue = "", property = "heylogs.tag.prefix")
+    @Parameter(property = "heylogs.tagPrefix")
     private String tagPrefix;
 
-    @Parameter(defaultValue = "", property = "heylogs.date")
+    @Parameter(property = "heylogs.date")
     private String date;
 
     @Override
     public void execute() throws MojoExecutionException {
-        if (skip) {
+        if (isSkip()) {
             getLog().info("Release has been skipped.");
             return;
         }
@@ -48,38 +47,32 @@ public final class ReleaseMojo extends HeylogsMojo {
             throw new MojoExecutionException("Changelog not found");
         }
 
-        release(loadVersion(), loadTagPrefix());
-    }
+        Document document = readChangelog(inputFile);
 
-    private Version loadVersion() throws MojoExecutionException {
-        return Version.of(ref, '-', DATE_PARSER.applyWithMojo(date));
-    }
-
-    private String loadTagPrefix() throws MojoExecutionException {
-        return TAG_PREFIX_PARSER.applyWithMojo(tagPrefix);
-    }
-
-    private void release(Version version, String tagPrefix) throws MojoExecutionException {
-        Heylogs heylogs = initHeylogs(false);
-
-        Document changelog = readChangelog(inputFile);
+        Version version = toVersion();
+        String tagPrefix = toTagPrefix();
 
         getLog().info("Releasing " + version + " with tag prefix '" + tagPrefix + "'");
-        heylogs.releaseChanges(changelog, version, tagPrefix);
+        initHeylogs(false).releaseChanges(document, version, tagPrefix);
 
-        writeChangelog(changelog, outputFile);
+        writeChangelog(document, inputFile);
     }
 
-    @MightBePromoted
-    private static String parseTagPrefix(String tagPrefix) {
+    @MojoParameterParsing
+    private @NonNull Version toVersion() throws MojoExecutionException {
+        return Version.of(ref, '-', parseLocalDate(date));
+    }
+
+    @MojoParameterParsing
+    private @NonNull String toTagPrefix() {
         return Objects.toString(tagPrefix, "");
     }
 
-    @MightBePromoted
-    private static LocalDate parseDate(String date) {
-        return date == null ? LocalDate.now(ZoneId.systemDefault()) : LocalDate.parse(date);
+    private static @NonNull LocalDate parseLocalDate(@Nullable String date) throws MojoExecutionException {
+        try {
+            return Version.parseLocalDate(date);
+        } catch (DateTimeParseException ex) {
+            throw new MojoExecutionException("Invalid format for 'date' parameter", ex);
+        }
     }
-
-    private static final MojoFunction<String, String> TAG_PREFIX_PARSER = of(ReleaseMojo::parseTagPrefix, "Invalid format for 'tagPrefix' parameter");
-    private static final MojoFunction<String, LocalDate> DATE_PARSER = of(ReleaseMojo::parseDate, "Invalid format for 'date' parameter");
 }