+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/ScanMojo.java
@@ -1,6 +1,7 @@
 package nbbrd.heylogs.maven.plugin;
 
-import internal.heylogs.StylishFormat;
+import nbbrd.console.picocli.MultiFileInputOptions;
+import nbbrd.console.picocli.text.TextOutputSupport;
 import nbbrd.heylogs.Heylogs;
 import nbbrd.heylogs.Scan;
 import org.apache.maven.plugin.MojoExecutionException;
@@ -11,51 +12,65 @@ import org.apache.maven.plugins.annotations.Parameter;
 import java.io.File;
 import java.io.IOException;
 import java.io.Writer;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
 
-import static internal.heylogs.maven.plugin.HeylogsParameters.*;
-import static java.util.Collections.singletonList;
+import static internal.heylogs.FormatSupport.resolveFormatId;
+import static internal.heylogs.HeylogsParameters.DEFAULT_CHANGELOG_FILE;
+import static internal.heylogs.HeylogsParameters.DEFAULT_RECURSIVE;
+import static java.util.stream.Collectors.toList;
+import static nbbrd.console.picocli.ByteOutputSupport.DEFAULT_STDOUT_FILE;
+import static nbbrd.console.picocli.text.TextOutputSupport.newTextOutputSupport;
 
+@lombok.Getter
+@lombok.Setter
 @Mojo(name = "scan", defaultPhase = LifecyclePhase.VALIDATE, threadSafe = true, requiresProject = false)
 public final class ScanMojo extends HeylogsMojo {
 
-    @Parameter(defaultValue = WORKING_DIR_CHANGELOG, property = INPUT_FILE_PROPERTY)
-    private File inputFile;
+    @Parameter(property = "heylogs.inputFiles", defaultValue = DEFAULT_CHANGELOG_FILE)
+    private List<File> inputFiles;
 
-    @Parameter(defaultValue = MOJO_LOG_FILE, property = OUTPUT_FILE_PROPERTY)
+    @Parameter(property = "heylogs.recursive", defaultValue = DEFAULT_RECURSIVE)
+    private boolean recursive;
+
+    @Parameter(property = "heylogs.outputFile", defaultValue = DEFAULT_STDOUT_FILE)
     private File outputFile;
 
-    @Parameter(defaultValue = StylishFormat.ID, property = FORMAT_ID_PROPERTY)
+    @Parameter(property = "heylogs.formatId")
     private String formatId;
 
     @Override
     public void execute() throws MojoExecutionException {
-        if (skip) {
+        if (isSkip()) {
             getLog().info("Scanning has been skipped.");
             return;
         }
 
-        if (inputFile.exists()) {
-            scan();
-        } else {
-            if (isRootProject()) {
-                raiseErrorMissingChangelog();
-            } else {
-                notifyMissingChangelog();
-            }
-        }
-    }
-
-    private void scan() throws MojoExecutionException {
         Heylogs heylogs = initHeylogs(false);
 
-        Scan scan = Scan
-                .builder()
-                .source(inputFile.toString())
-                .summary(heylogs.scanContent(readChangelog(inputFile)))
-                .build();
+        List<Scan> list = new ArrayList<>();
+        try {
+            MultiFileInputOptions input = new MultiFileInputOptions();
+            input.setFiles(inputFiles.stream().map(File::toPath).collect(toList()));
+            input.setRecursive(recursive);
+            input.setSkipErrors(false);
+            for (Path file : input.getAllFiles(HeylogsMojo::accept)) {
+                list.add(Scan
+                        .builder()
+                        .source(file.toString())
+                        .summary(heylogs.scanContent(readChangelog(file.toFile())))
+                        .build());
+            }
+        } catch (IOException ex) {
+            throw new MojoExecutionException("Error while listing files", ex);
+        }
+
+        TextOutputSupport outputSupport = newTextOutputSupport();
+        String formatId = resolveFormatId(getFormatId(), heylogs, outputSupport::isStdoutFile, outputFile.toPath());
 
         try (Writer writer = newWriter(outputFile, getLog()::info)) {
-            heylogs.formatStatus(formatId, writer, singletonList(scan));
+            heylogs.formatStatus(formatId, writer, list);
         } catch (IOException ex) {
             throw new MojoExecutionException("Error while writing", ex);
         }